refactor(): restructure source files within subdirectories
