+++ src/main/kotlin/ecommerce/config/GlobalExceptionHandler.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.config
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
