+++ src/main/kotlin/ecommerce/controller/ProductPageController.kt
@@ -1,5 +1,6 @@
-package ecommerce
+package ecommerce.controller
 
+import ecommerce.service.ProductService
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
