+++ src/main/kotlin/ecommerce/repository/ProductRepository.kt
@@ -1,5 +1,6 @@
-package ecommerce
+package ecommerce.repository
 
+import ecommerce.entity.Product
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.RowMapper
 import org.springframework.stereotype.Repository
