+++ src/test/kotlin/ecommerce/ProductTest.kt
@@ -1,5 +1,6 @@
 package ecommerce
 
+import ecommerce.entity.Product
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Test
